handle different name for assignment_id and filename

+++ nbgrader/exchange/collect.py
@@ -28,8 +28,8 @@ class ExchangeCollect(Exchange):
     def _path_to_record(self, path):
         filename = os.path.split(path)[1]
         # Only split twice on +, giving three components. This allows usernames with +.
-        filename_list = filename.rsplit('+', 3)
-        if len(filename_list) < 3:
+        filename_list = filename.rsplit('+', 2)
+        if len(filename_list) != 3:
             self.fail("Invalid filename: {}".format(filename))
         username = filename_list[0]
         timestamp = parse_utc(filename_list[2])
@@ -67,7 +67,7 @@ class ExchangeCollect(Exchange):
                 len(self.src_records),
                 self.coursedir.assignment_id,
                 self.course_id))
-
+                
         user_infos = []
         for rec in self.src_records:
             student_id = rec['username']
@@ -96,7 +96,8 @@ class ExchangeCollect(Exchange):
                     self.log.info("Updating submission: {} {}".format(student_id, self.coursedir.assignment_id))
                     shutil.rmtree(dest_path)
                     self.log.info("Updating hashed_submission directory")
-                    shutil.rmtree(hashed_dest_path)
+                    if os.path.isdir(hashed_dest_path):
+                        shutil.rmtree(hashed_dest_path)
                 else:
                     self.log.info("Collecting submission: {} {}".format(student_id, self.coursedir.assignment_id))
                 self.do_copy(src_path, dest_path)
@@ -115,31 +116,35 @@ class ExchangeCollect(Exchange):
             user_info = self.extract_user_info(dest_path, student_id)
             user_infos.append(user_info)
 
-        #Create hashcode list            
-        self.log.info("Creating hashcode list")
-        csv_filename = "hashcode_list.csv"
-        html_filename = "hashcode_list.html"
-        hashcode_list_path = os.path.join(self.coursedir.root, self.coursedir.submitted_directory)
-        csv_file_dest_path = os.path.join(hashcode_list_path, csv_filename)
-        html_file_dest_path = os.path.join(hashcode_list_path, html_filename)
-        self.create_hashcode_list(user_infos, hashcode_list_path, filename=csv_filename)
+        #Create hashcode list
+        if user_info is not None: 
+            self.log.info("Creating hashcode list")
+            csv_filename = "hashcode_list.csv"
+            html_filename = "hashcode_list.html"
+            hashcode_list_path = os.path.join(self.coursedir.root, self.coursedir.submitted_directory)
+            csv_file_dest_path = os.path.join(hashcode_list_path, csv_filename)
+            html_file_dest_path = os.path.join(hashcode_list_path, html_filename)
+            self.create_hashcode_list(user_infos, hashcode_list_path, filename=csv_filename)
 
-        # Load csv and create html
-        self.log.info("Creating {}".format(html_file_dest_path))
-        data = pd.read_csv(csv_file_dest_path) 
-        data.to_html(html_file_dest_path, justify='center')
+            # Load csv and create html
+            self.log.info("Creating {}".format(html_file_dest_path))
+            data = pd.read_csv(csv_file_dest_path) 
+            data.to_html(html_file_dest_path, justify='center')
 
     def extract_user_info(self, info_path, username):
         file_info_path = os.path.join(info_path, username+"_info.txt")
-        try: 
-            with open(file_info_path) as f:
-                user_info = f.readlines()
-        finally:  
-            f.close()
+        if os.path.isfile(file_info_path):
+            try: 
+                with open(file_info_path) as f:
+                    user_info = f.readlines()
+            finally:  
+                f.close()
 
-        user_info = [x.strip().split(': ')[1] for x in user_info]
+            user_info = [x.strip().split(': ')[1] for x in user_info]
 
-        return user_info
+            return user_info
+        else:
+            return
 
     def create_hashcode_list(self, user_infos, list_dest_path, filename='hashcode_list.csv'):
         username_field = "Username"
